move test functions to tests/utilities/test_tasks
